feat(key_switch): 优化按键初始化流程，增加freedorm_button_init
