+++ ble_detect_hid/components/bsp_button/button.h
@@ -1,13 +1,16 @@
 #ifndef BUTTON_H
 #define BUTTON_H
 
-#include <stdlib.h>
+#include "button.h"
+#include "esp_log.h"
+#include "driver/gpio.h"
+#include "freertos/FreeRTOS.h"
+#include "freertos/task.h"
+#include "freertos/event_groups.h"
+#include "multi_button.h"
+#include "esp_mac.h"
 
 #define PAIRING_BUTTON_GPIO GPIO_NUM_0 // 修改为您的按键GPIO编号
-#define OUTPUT_LED_D4 GPIO_NUM_12
-#define OUTPUT_LED_D5 GPIO_NUM_13
-#define CTL_LOCK GPIO_NUM_6
-#define CTL_D0 GPIO_NUM_3
 
 typedef enum
 {
@@ -23,6 +26,7 @@ extern uint32_t led_state_mask; // 在这里初始化，位图，记录每个 GP
 // 函数声明
 
 uint8_t read_button_GPIO(uint8_t button_id);
+void freedorm_button_init();
 void button_task(void *arg);
 
 void BTN1_PRESS_REPEAT_Handler(void *btn);
