+++ ble_detect_hid/main/_freedorm_main.c
@@ -22,6 +22,7 @@
 #include "button.h"
 #include "ble_module.h"
 #include "ws2812b_led.h"
+#include "lock_control.h"
 
 /**
  * Brief:
@@ -46,8 +47,6 @@
 
 #define BUTTON1_ID 0
 
-struct Button btn1;
-
 void app_main(void)
 {
     esp_err_t ret;
@@ -61,20 +60,10 @@ void app_main(void)
     }
     ESP_ERROR_CHECK(ret);
     // 按键初始化
-    uint64_t gpio_intput_sel = (1ULL << PAIRING_BUTTON_GPIO);
     uint64_t gpio_output_sel = (1ULL << OUTPUT_LED_D4) | (1ULL << OUTPUT_LED_D5) | (1ULL << CTL_LOCK) | (1ULL << CTL_D0);
 
-    gpio_config_t input_io_conf = {
-        .pin_bit_mask = gpio_intput_sel,      // 配置 GPIO0 为输入
-        .mode = GPIO_MODE_INPUT,              // 输入模式
-        .pull_up_en = GPIO_PULLUP_DISABLE,    // 上拉
-        .pull_down_en = GPIO_PULLDOWN_ENABLE, // 不需要下拉
-        .intr_type = GPIO_INTR_DISABLE        // 不需要中断
-    };
-    gpio_config(&input_io_conf);
-
     gpio_config_t output_io_conf = {
-        .pin_bit_mask = gpio_output_sel,       // 配置 GPIO2 为输出
+        .pin_bit_mask = gpio_output_sel,
         .mode = GPIO_MODE_OUTPUT,              // 输出模式
         .pull_up_en = GPIO_PULLUP_DISABLE,     // 不需要上拉
         .pull_down_en = GPIO_PULLDOWN_DISABLE, // 不需要下拉
@@ -82,17 +71,9 @@ void app_main(void)
     };
     gpio_config(&output_io_conf);
 
-    // 初始化按键对象
-    button_init(&btn1, read_button_GPIO, 1, 0); // 第三个参数为有效电平 0（低电平有效），第四个参数为按键 ID 艹，tmd debug半天结果是这里参考电平的问题，艹
-
-    button_attach(&btn1, PRESS_REPEAT, BTN1_PRESS_REPEAT_Handler);
-    button_attach(&btn1, SINGLE_CLICK, BTN1_SINGLE_CLICK_Handler);
-    button_attach(&btn1, DOUBLE_CLICK, BTN1_DOUBLE_CLICK_Handler);
-    button_attach(&btn1, LONG_PRESS_START, BTN1_LONG_PRESS_START_Handler);
-    button_attach(&btn1, LONG_PRESS_HOLD, BTN1_LONG_PRESS_HOLD_Handler);
-    button_start(&btn1);
-
     ble_module_init();
     ws2812b_led_init(); // 按键在之后初始化，因为按键依赖ws2812b的队列
+    freedorm_button_init();
+    lock_control_init();
     xTaskCreate(&button_task, "button_task", 2048, NULL, 1, NULL);
 }
